Corrección de inconsistencias del backend de los controllers, modelos y rutas de users,cards,tags,admin,validaciones y autenticación y vistas del frontend, aplicación en local 100% funcional

+++ frontend/src/services/userService.js
@@ -2,7 +2,7 @@ import api from "../services/api";
 
 //Obtener los usuarios (admin)
 export const getUsers = async() => {
-    const res = await api.get("/users");
+    const res = await api.get("/admin/users");
     return res.data;
 };
 
@@ -20,18 +20,18 @@ export const updateProfile = async (data) => {
 
 //Crear usuario con rol administrador
 export const createUser = async (data) => {
-    const res = await api.post("/users", data);
+    const res = await api.post("/admin/users", data);
     return res.data;
 };
 
 //Actualizar usuario con rol administrador
-export const updateUser = async (id,data) => {
-    const res = await api.put(`/users/${id}`, data);
+export const updateUser = async (id, data) => {
+    const res = await api.put(`/admin/users/${id}`, data);
     return res.data;
 };
 
 //Eliminar usuario con rol administrador
 export const deleteUser = async (id) => {
-    const res = await api.delete(`/users/${id}`);
+    const res = await api.delete(`/admin/users/${id}`);
     return res.data;
 };
